refactor: 코드 리뷰 최소 반영 및 기존 레이싱 관련 클래스 v1 폴더로 이관한다
