+++ README.md
@@ -118,15 +118,15 @@
 
 #### [Todo List]
 
-- [ ] 피드백을 반영한다.
-- [ ] 린트를 체크한다.
-- [ ] 자동차는 이름을 가진다.
-- [ ] 자동차의 이름은 5글자를 초과할 수 없다.
+- [x] 피드백을 반영한다.
+- [ ] 자동차는 0부터 9까지의 임의의 수 중에 4 이상일 경우에만 전진할 수 있다.
+- [ ] 자동차는 위치를 가지고, 기본 값은 0이다.
+- [ ] 자동차는 이름을 가지고, 5글자를 초과할 수 없다.
 - [ ] 자동차 이름들은 `,` 기준으로 구분한다.
 - [ ] 게임 완료 후 우승자가 1명일 때 해당 이름을 출력한다.
 - [ ] 게임 완료 후 우승자가 여러명일 때 해당 이름들을 출력한다.
 - [ ] 모든 indent depth는 2를 넘지 않는다.
 - [ ] 모든 함수 길이는 15라인 이하로 제한한다.
-
+- [ ] 린트를 체크한다.
