--- src/main/kotlin/racing/RandomNumberGenerator.kt
@@ -1,9 +1,0 @@
-package racing
-
-class RandomNumberGenerator {
-    companion object {
-        fun generate(from: Int, to: Int): Int {
-            return (from..to).random()
-        }
-    }
-}
