--- src/main/kotlin/racing/ResultView.kt
@@ -1,12 +1,0 @@
-package racing
-
-class ResultView {
-    companion object {
-        fun printPosition(position: Int) {
-            repeat(position) {
-                print('-')
-            }
-            println()
-        }
-    }
-}
