+++ src/main/kotlin/racing/v1/Car.kt
@@ -1,7 +1,8 @@
-package racing
+package racing.v1
 
 class Car {
     var position: Int = 0
+        private set
 
     fun moveForward(number: Int) {
         if (DrivingPolicy.canMoveForward(number)) {
