+++ src/main/kotlin/racing/v1/DrivingPolicy.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.v1
 
 class DrivingPolicy {
     companion object {
