+++ src/main/kotlin/racing/v1/InputView.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.v1
 
 class InputView {
     companion object {
