+++ src/main/kotlin/racing/v1/ResultView.kt
@@ -1,0 +1,11 @@
+package racing.v1
+
+object ResultView {
+    fun printPosition(position: Int) {
+        repeat(position) {
+            print('-')
+        }
+        println()
+    }
+}
+
